refactor(vendors,auth): rename firebase dir and enforce adapter boundary
