+++ package.json
@@ -23,8 +23,8 @@
     "preview": "vite preview",
     "preview:https": "PREVIEW_HTTPS=1 vite preview",
     "preview:pwa": "pnpm build && pnpm preview:https",
-    "lint": "pnpm lint:boundaries",
-    "lint:boundaries": "eslint \"src/**/*.js\"",
+    "lint": "pnpm lint:boundaries && eslint \"src/**/*.js\"",
+    "lint:boundaries": "eslint --config eslint.boundaries.config.js \"src/**/*.js\"",
     "lint:b:all": "BOUNDARIES_ENFORCE_ALL=1 pnpm lint:boundaries",
     "test": "vitest --run",
     "test:node": "vitest --run --project node",
