+++ src/features/call/tests/member-listener-fresh-join.test.js
@@ -75,7 +75,7 @@ vi.mock('firebase/database', () => {
 });
 
 // Provide a minimal app export for rtdb initialization
-vi.mock('../../../firebase/firebase', () => ({ app: {} }));
+vi.mock('../../../vendors/firebase', () => ({ app: {} }));
 
 // Mock the diagnostic logger
 vi.mock('../../../utils/dev/diagnostic-logger.js', () => ({
