+++ src/vendors/firebase.js
